requirements.txt
# Web Framework & API
streamlit
Flask
flask-cors
Werkzeug

# AI & LLM Core
google-generativeai
langchain
langchain-core
langchain-classic
langchain-text-splitters
langchain_google_genai
langchain-community

# Data Processing & Vector DB
PyPDF2
faiss-cpu

# Utilities
python-dotenv
pyyaml
requests
urllib3

# Language & Translation
langdetect
deep-translator

# Google OAuth
google-auth
google-auth-oauthlib
google-auth-httplib2